Move files to the wallet folders
